chore(i dont know anymore): deadline rushing

+++ digimon/digimon/routers/transaction.py
@@ -1,10 +1,7 @@
 from fastapi import APIRouter, HTTPException, Depends, Query
-
 from typing import Optional, Annotated
-
 from sqlmodel import Field, SQLModel, Session, select, func
 from sqlmodel.ext.asyncio.session import AsyncSession
-
 import math
 
 from .. import models
@@ -14,31 +11,27 @@ router = APIRouter(prefix="/transactions" , tags=["transactions"])
 
 SIZE_PER_PAGE = 50
 
-
 @router.get("")
 async def read_transactions(
     session: Annotated[AsyncSession, Depends(models.get_session)],
     page: int = 1,
 ) -> models.TransactionList:
+    if page < 1:
+        page = 1
 
     result = await session.exec(
-        select(models.DBTransaction).offset((page - 1) * SIZE_PER_PAGE).limit(SIZE_PER_PAGE)
+        select(models.DBTransaction)
+        .offset((page - 1) * SIZE_PER_PAGE)
+        .limit(SIZE_PER_PAGE)
     )
     transactions = result.all()
 
-    page_count = int(
-        math.ceil(
-            (await session.exec(select(func.count(models.DBTransaction.id)))).first()
-            / SIZE_PER_PAGE
-        )
-    )
+    total_count = await session.exec(select(func.count(models.DBTransaction.id)))
+    page_count = int(math.ceil(total_count.first() / SIZE_PER_PAGE))
 
-    print("page_count", page_count)
-    print("transactions", transactions)
     return models.TransactionList.from_orm(
         dict(transactions=transactions, page_count=page_count, page=page, size_per_page=SIZE_PER_PAGE)
     )
-
 
 @router.post("")
 async def create_transaction(
@@ -46,27 +39,22 @@ async def create_transaction(
     current_user: Annotated[models.User, Depends(deps.get_current_user)],
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> models.Transaction | None:
-    # Query to find the wallet based on the foreign key relationship
     wallet = await session.get(models.Wallet, current_user.id)
     if not wallet:
         raise HTTPException(status_code=404, detail="Wallet not found")
 
-    # Query to find the item based on the foreign key relationship
     item = await session.get(models.Item, transaction.item_id)
     if not item:
         raise HTTPException(status_code=404, detail="Item not found")
 
-    # Check if the wallet balance is sufficient
     if wallet.balance < item.price:
         raise HTTPException(status_code=400, detail="Not enough money")
 
-    # Update the wallet balance
     wallet.balance -= item.price
     session.add(wallet)
     await session.commit()
     await session.refresh(wallet)
 
-    # Create and add the transaction
     db_transaction = models.DBTransaction(
         amount=transaction.amount,
         item_id=transaction.item_id,
@@ -76,9 +64,6 @@ async def create_transaction(
     await session.commit()
     await session.refresh(db_transaction)
     return models.Transaction.from_orm(db_transaction)
-
-   
-
 
 @router.get("/{transaction_id}")
 async def read_transaction(
@@ -97,7 +82,9 @@ async def delete_transaction(
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> dict:
     db_transaction = await session.get(models.DBTransaction, transaction_id)
-    await session.delete(db_transaction)
-    await session.commit()
-
-    return dict(message="delete success")
+    if db_transaction:
+        await session.delete(db_transaction)
+        await session.commit()
+        return dict(message="delete success")
+    else:
+        raise HTTPException(status_code=404, detail="Transaction not found")
